Make subnavigation router paths dynamic
