--- src/scripts/components/screens/DemoHome.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function DemoHome() {
-  return <a href="/wizard">Top-Level Wizard Demo</a>;
-}
